requirements.txt
openai-agents>=0.0.19
openai>=1.0.0
ollama>=0.1.0
httpx>=0.24.0
pydantic>=2.0.0
python-dotenv>=1.0.0
aiofiles>=23.0.0
typing-extensions>=4.8.0